requirements.txt
asttokens==3.0.0
attrs==24.3.0
Automat==24.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect==1.2.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
filelock==3.16.1
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==8.31.0
itemadapter==0.10.0
itemloaders==1.3.2
jedi==0.19.2
jmespath==1.0.1
lxml==5.3.0
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==24.2
parsel==1.9.1
parso==0.8.4
playwright==1.49.1
prompt_toolkit==3.0.48
Protego==0.3.1
proto-plus==1.25.0
protobuf==5.29.2
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyee==12.0.0
Pygments==2.18.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-dotenv==1.0.1
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9
Scrapy==2.12.0
scrapy-playwright==0.0.42
service-identity==24.2.0
setuptools==75.6.0
stack-data==0.6.3
tldextract==5.1.3
traitlets==5.14.3
Twisted==24.11.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
w3lib==2.2.1
wcwidth==0.2.13
zope.interface==7.2
